requirements.txt
canvasapi
requests
notebooklm-py
playwright
python-dotenv
